Add files via upload
Class exercise using a while loop with a counter to determine when the while loop is finished.  This program collects 7 days of bottles for return for deposit, calculates the total payout for the returned bottles and prints out the total number of bottles returned and the amount paid out.  If offers to continue entering data for another 7 day period.  This decision is controlled by a boolean condition set to True or False based on the input from the user.